Finished Lab 4
Once I figured out the way that I wanted to layout the controls it wasn't too bad. Got stuck for a while getting the split view to appear on top of the images instead of below it; I had to dig around in the functions for a long time to find the command for it.