Not finished with Lab 6
Looked online to find how the forecast endpoints work. Figured out how to iterate through the responses to populate the observable collection, but still can't get the information out of the responses. All the data I need in the responses is in responses.period[#], but I can't figure out what the index I'm supposed to put in for the period.