Finished Researching API's, Planning out App
Got checked off on the scope of the project. I'm going to have a song database, where users can search for songs and have all the information on the song compiled on a page. It will include lyrics of the song, information about the song pulled from wikipedia, and a music video that I'm trying to pull from YouTube.