Need to find new API for music
After talking to Mr. Smith, found out that playing music directly in the app would be difficult, and that figuring out how to deal with the oauth code needed for the API would be too complicated. I checked other API's, and the other music ones also need oauth, so my app may not be able to play music. I'm researching other API's, and if anything I want to include a link to a website where the user can play the music on thier own.