Working on Creating Models
Made the general layout of the app and working on adding in the csharp code from json. Running into an issue with the MediaWiki API where some names and types are invalid. I can just change the names of the classes to something valid, but I do not know how to fix invalid types. Am looking for a solution or another endpoint without these problems.