Working on Converting Root Objects from API calls into Usable View Models
Using Lab 6 as a template. Finished the retrieval methods. Made a new view model for music, which will be used in a list on the results page. Having an issue with visibility on the music search result view model. I'm looking back at lab 6 and making sure everythings the same.