Finished Functionality on Results Page, Working on Song Page
Songs are being properly displayed on the results screen and clicking on one correctly jumps to the song page. There is an issue on the song page when it opens, where the api calls to get music and get lyrics are not working. Need to check the retrieval class and make sure I'm using the api's common id vs id correctly in my calls.